glitchy rendering logic

+++ src/cell.cxx
@@ -1,12 +1,14 @@
 // MIT License
+
 #include "cell.h"
 #include <iostream>
 
-Cell::Cell(int x1, int y1) {
+Cell::Cell(int64_t x1, int64_t y1) {
 	this->x = x1;
 	this->y = y1;
 }
 
+/*
 bool Cell::addCell(Cell) {
 
 }
@@ -18,3 +20,4 @@ bool Cell::removeCell(Cell) {
 void Cell::iterateCell() {
 
 }
+*/
